modifying acc to rest of the stack
