+++ atk.env
@@ -2,5 +2,5 @@ DOCKERHUB_USERNAME=uwsbel
 
 ROS_DISTRO=humble
 
-DEFAULT_APT_DEPENDENCIES="bash vim git git-lfs python3-pip ros-humble-robot-localization"
+DEFAULT_APT_DEPENDENCIES="bash vim git git-lfs python3-pip"
 DEFAULT_PIP_REQUIREMENTS=""
