+++ atk.yml
@@ -36,7 +36,7 @@ services:
         ROS_DISTRO: "${ROS_DISTRO}"
         USER_UID: "${USER_UID:-1000}"
         USER_GID: "${USER_GID:-1000}"
-        APT_DEPENDENCIES: "${DEFAULT_APT_DEPENDENCIES:-}"
+        APT_DEPENDENCIES: "${DEFAULT_APT_DEPENDENCIES:-} ros-${ROS_DISTRO}-robot-localization"
         PIP_REQUIREMENTS: "${DEFAULT_PIP_REQUIREMENTS:-}"
     working_dir: "/home/${COMPOSE_PROJECT_NAME}/${COMPOSE_PROJECT_NAME}"
     volumes:
@@ -73,7 +73,7 @@ services:
         IMAGE_BASE: "dustynv/ros"
         IMAGE_TAG: "humble-pytorch-l4t-r35.3.1"
         ROS_INSTALL_PREFIX: "/opt/ros/${ROS_DISTRO}/install"
-        APT_DEPENDENCIES: "${DEFAULT_APT_DEPENDENCIES}"
+        APT_DEPENDENCIES: "${DEFAULT_APT_DEPENDENCIES} ros-${ROS_DISTRO}-robot-localization"
         PIP_REQUIREMENTS: "${DEFAULT_PIP_REQUIREMENTS}"
         ROSDEP_SKIP_KEYS: "python3-torchvision tf_transformations"
   art5:
