+++ workspace/src/common/launch/art_localization_launch/launch/ekf_launch.launch.py
@@ -1,0 +1,95 @@
+#
+# BSD 3-Clause License
+#
+# Copyright (c) 2022 University of Wisconsin - Madison
+# All rights reserved.
+#
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that the following conditions are met:
+#
+# * Redistributions of source code must retain the above copyright notice, this
+#   list of conditions and the following disclaimer.
+#
+# * Redistributions in binary form must reproduce the above copyright notice,
+#   this list of conditions and the following disclaimer in the documentation
+#   and/or other materials provided with the distribution.
+#
+# * Neither the name of the copyright holder nor the names of its
+#   contributors may be used to endorse or promote products derived from
+#   this software without specific prior written permission.
+#
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+# AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+# IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+# DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
+# FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+# DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+# SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+# CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+# OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+# OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.#
+
+# ros imports
+from launch import LaunchDescription
+from launch_ros.actions import Node
+
+# internal imports
+from launch_utils import AddLaunchArgument, GetLaunchArgument, GetPackageSharePath
+from enum import Enum
+
+
+def generate_launch_description():
+    ekf_config_path = GetPackageSharePath(
+        package="art_localization_launch", path="config/ekf_config.yaml"
+    )
+    # Transforms
+    gps_transform = Node(
+        package="tf2_ros",
+        executable="static_transform_publisher",
+        name="gps_transform",
+        arguments=["0", "0", "0", "0", "0", "0", "base_link", "gps"],
+    )
+    imu_transform = Node(
+        package="tf2_ros",
+        executable="static_transform_publisher",
+        name="imu_transform",
+        arguments=["0", "0", "0", "3.1416", "-1.5708", "0", "base_link", "imu"],
+    )
+
+    # Navsat Transform
+    navsat_transform_node = Node(
+        package="robot_localization",
+        executable="navsat_transform_node",
+        name="navsat_transform_node",
+        respawn=True,
+        parameters=[
+            {
+                "magnetic_declination_radians": 0.0,
+                "yaw_offset": 0.0,
+                "zero_altitude": True,
+                "use_odometry_yaw": True,
+                "wait_for_datum": False,
+                "publish_filtered_gps": True,
+                "broadcast_cartesian_transform": False,
+            }
+        ],
+        remappings=[
+            ("/imu/data", "/imu"),
+            ("/gps/fix", "/fix"),
+            # ('/odometry/filtered', '/odometry/gps')
+        ],
+    )
+
+    ekf_filter_node = Node(
+        package="robot_localization",
+        executable="ekf_node",
+        name="ekf_filter_node",
+        output="screen",
+        parameters=[
+            ekf_config_path,
+        ],
+    )
+
+    return LaunchDescription(
+        [gps_transform, imu_transform, navsat_transform_node, ekf_filter_node]
+    )
