+++ workspace/src/localization/localization_shared_utils/setup.py
@@ -1,15 +1,6 @@
 from setuptools import setup
-import os
 
 package_name = "localization_shared_utils"
-
-
-launch_directory = os.path.join("launch")
-launch_files = [
-    os.path.join(launch_directory, file)
-    for file in os.listdir(launch_directory)
-    if file.endswith(".py")
-]
 
 setup(
     name=package_name,
@@ -18,7 +9,6 @@ setup(
     data_files=[
         ("share/ament_index/resource_index/packages", ["resource/" + package_name]),
         ("share/" + package_name, ["package.xml"]),
-        (os.path.join("share", package_name, "launch"), launch_files),
     ],
     install_requires=["setuptools"],
     zip_safe=True,
